requirements.txt
asgiref==3.2.7
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.6
django-environ==0.4.5
django-storages==1.9.1
google-api-core==1.17.0
google-auth==1.14.3
google-cloud-core==1.3.0
google-cloud-secret-manager==0.2.0
google-cloud-storage==1.28.1
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpc-google-iam-v1==0.12.3
grpcio==1.28.1
gunicorn==20.0.4
idna==2.9
protobuf==3.11.3
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
requests==2.23.0
rsa==4.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
